==> scripts/0_Install_R_packages_and_check.Rmd <==
---
title: "R Notebook - 1_Install_R_packages_and_check"
output: html_notebook
---
#### Install required Packages listed below from Bioconductor (need to run once in a device) 
```{r chunk 1}
if (!require("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
BiocManager::install("XVector",force = TRUE)
BiocManager::install("multtest")
BiocManager::install("glmGamPoi")
```

#### Install required Packages listed below from CRAN (need to run once in a device) 
```{r chunk 2}
install.packages('tidyverse')
install.packages('Matrix')
install.packages('RCurl')
install.packages('scales')
install.packages('metap')
install.packages('Seurat')
install.packages("ggplot2")
install.packages("sctransform")
```

##### Check for successful installation by loading the packages
```{r chunk 3}
library(XVector)
library(Seurat)
library(tidyverse)
library(Matrix)
library(RCurl)
library(scales)
library(sctransform)
```
------End------
